fixed background color to show up on all pages also moved donation from out of components and into pages and moved client id
