+++ client/package.json
@@ -5,14 +5,11 @@
   "dependencies": {
     "@apollo/client": "^3.6.1",
     "@apollo/react-hooks": "^4.0.0",
-<<<<<<< HEAD
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/free-regular-svg-icons": "^6.1.1",
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
     "@fortawesome/react-fontawesome": "^0.1.18",
-=======
     "@paypal/react-paypal-js": "^7.6.0",
->>>>>>> d3cabd825b55be457efb70569d18bb2a1a15c611
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.1.1",
     "@testing-library/user-event": "^13.5.0",
