+++ client/src/App.js
@@ -3,7 +3,7 @@ import Header from "./components/Header";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Homepage from "./pages/Homepage";
-import MyParks from "./pages/MyParks"
+import MyParks from "./pages/MyParks";
 import LoginForm from "./pages/Login";
 import SignupForm from "./pages/Signup";
 import {
@@ -11,20 +11,20 @@ import {
   InMemoryCache,
   ApolloProvider,
   createHttpLink,
-} from '@apollo/client';
-import { setContext } from '@apollo/client/link/context';
-import Donation from "./components/PayPal";
+} from "@apollo/client";
+import { setContext } from "@apollo/client/link/context";
+import Donation from "./pages/Donation";
 
 const httpLink = createHttpLink({
-  uri: '/graphql',
+  uri: "/graphql",
 });
 
 const authLink = setContext((_, { headers }) => {
-  const token = localStorage.getItem('id_token');
+  const token = localStorage.getItem("id_token");
   return {
     headers: {
       ...headers,
-      authorization: token ? `Bearer ${token}` : '',
+      authorization: token ? `Bearer ${token}` : "",
     },
   };
 });
@@ -34,24 +34,22 @@ const client = new ApolloClient({
   cache: new InMemoryCache(),
 });
 
-
 function App() {
   return (
     <ApolloProvider client={client}>
-    <Router>
-      <>
-        <Header />
-        <Routes>
-          <Route path="/" element={<Homepage />} />
-          <Route path="/myparks" element={<MyParks />} />
-          <Route path="/login" element={<LoginForm />} />
-          <Route path="/Signup" element={<SignupForm />} />
-          <Route path="/donation" element={<Donation />} />
-          {/* <Route path="/myparks:id" element={<MyParks />} /> REPLACE WITH :ID VERSION ONCE GRAPHQL IN PLACE*/}
-        </Routes>
-        
-      </>
-    </Router>
+      <Router>
+        <>
+          <Header />
+          <Routes>
+            <Route path="/" element={<Homepage />} />
+            <Route path="/myparks" element={<MyParks />} />
+            <Route path="/login" element={<LoginForm />} />
+            <Route path="/Signup" element={<SignupForm />} />
+            <Route path="/donation" element={<Donation />} />
+            {/* <Route path="/myparks:id" element={<MyParks />} /> REPLACE WITH :ID VERSION ONCE GRAPHQL IN PLACE*/}
+          </Routes>
+        </>
+      </Router>
     </ApolloProvider>
   );
 }
